RISC-V: drivers/iommu: Optimize invalidations after map/unmap
When map_pages() and unmap_pages() are invoked by the Linux common IOMMU
code, they are followed by invocation of iotlb_sync_map() and
iotlb_sync() respectively. These functions are meant for invalidating
only the range covered by the mapped or unmapped pages.
That's why instead of invalidating all IOATC entries along with all
DevATC entries, they are updated to allow for fine grain handling.

That means mapping or unmapping a page range won't discard entries that
are not part of this range in the different caches. This will imply a
performance gain for applications that are not optimized to reuse
existing mappings.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>